requirements.txt
python-slugify==3.0.1
redis==3.2.1
celery==4.3.0rc3
django-celery-email==2.0.1
daphne==2.2.5
django==2.1.7
django-redis==4.10.0
django-allauth==0.39.1
django-environ==0.4.5
django-crispy-forms==1.7.2
django-compressor==2.2
mysqlclient==1.4.2.post1
django-contrib-comments==1.9.1
django-markdownx==2.0.28
channels==2.1.7
sorl-thumbnail==12.5.0
django-taggit==1.1.0
channels-redis==2.3.3
awesome-slugify==1.6.5
argon2-cffi==19.1.0
pillow==5.4.1
python3-openid==3.1.0
requests-oauthlib==1.2.0
requests==2.21.0
django-haystack==2.8.1
elasticsearch==2.4.1
django-debug-toolbar==2.2
django_extensions==3.1.0
uwsgi==2.0.18
